pyproject.toml
[project]
name = "facility-intelligence-system"
version = "0.1.0"
description = "AI-powered facility monitoring and intelligence system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite",
    "pydantic>=2.12",
    "langgraph>=1.0",
    "langchain>=1.2.7",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.0",
    "langchain-anthropic>=1.0",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=1.0",
    "httpx>=0.27",
    "ruff",
    "poethepoet>=0.24",
]

[tool.setuptools.packages.find]
where = ["backend"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]
ignore = [
    "B008",  # Allow Depends() in function defaults (FastAPI pattern)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]

[tool.poe.tasks]
backend = { shell = "cd backend && uvicorn app.main:app --reload --port 8000" }
frontend = { shell = "cd frontend && npm run dev" }
frontend-mock = { shell = "cd frontend && VITE_USE_MOCK=true npm run dev" }
db-setup = { shell = "cd backend && python scripts/setup_database.py" }
test = { shell = "cd backend && pytest" }
lint = { shell = "ruff check backend" }
format = { shell = "ruff format backend" }
